Incorporate feedbacks
