+++ ui/src/__tests__/_horizontal-possibility-panel.test.js
@@ -5,9 +5,18 @@ import HorizontalPossibilityPanel from "../pages/_horizontal-possibility-panel";
 
 describe("HorizontalPossibilityPanel", () => {
   const mockQueries = [
-    "What's the weather like tomorrow?",
-    "Show me vacation ideas.",
-    "Plan a weekend getaway.",
+    {
+      name: "What is the weather today?",
+      description: "What is the weather today?",
+    },
+    {
+      name: "Show me vacation ideas.",
+      description: "Show me vacation ideas.",
+    },
+    {
+      name: "Plan a weekend getaway.",
+      description: "Plan a weekend getaway.",
+    },
   ];
 
   it("renders the suggestions", () => {
@@ -23,7 +32,7 @@ describe("HorizontalPossibilityPanel", () => {
     expect(screen.getByText("Suggestions:")).toBeInTheDocument();
 
     mockQueries.forEach((query) => {
-      expect(screen.getByText(query)).toBeInTheDocument();
+      expect(screen.getByText(query.description)).toBeInTheDocument();
     });
   });
 
@@ -35,11 +44,11 @@ describe("HorizontalPossibilityPanel", () => {
         onClick={mockOnClick}
       />,
     );
-    const firstButton = screen.getByText(mockQueries[0]);
+    const firstButton = screen.getByText(mockQueries[0].description);
 
     fireEvent.click(firstButton);
 
     expect(mockOnClick).toHaveBeenCalledOnce();
-    expect(mockOnClick).toHaveBeenCalledWith(mockQueries[0]);
+    expect(mockOnClick).toHaveBeenCalledWith(mockQueries[0].description);
   });
 });
