+++ ui/src/pages/_cards-chat.js
@@ -25,8 +25,7 @@ import {
 import PromptPreview from "../app/_prompt_preview";
 import MarkdownRenderer from "../app/_markdown_renderer";
 import { scenarioToText } from "../app/_dynamic_data_renderer";
-import HorizontalPossibilityPanel from "./_horizontal-possibility-panel";
-import Disclaimer from "../pages/_disclaimer";
+import EnrichCard from "../app/_encrich_card";
 
 const CardsChat = ({
   promptId,
@@ -46,7 +45,6 @@ const CardsChat = ({
   const [selectedContexts, setSelectedContexts] = useState([]);
   const [promptInput, setPromptInput] = useState("");
 
-  const [iterationPrompt, setIterationPrompt] = useState("");
   const [enableGenerateMoreCards, setEnableGenerateMoreCards] = useState(true);
 
   const [cardExplorationDrawerOpen, setCardExplorationDrawerOpen] =
@@ -289,130 +287,6 @@ const CardsChat = ({
   const sendGetMorePrompt = () => {
     sendCardBuildingPrompt(buildRequestDataGetMore());
   };
-
-  /** ITERATION EXPERIMENT
-   * (behind feature toggle, experimental implementation)
-   *
-   * Switch on with:
-   * window.localStorage.setItem("toggles", '{ "cards_iteration": true }')
-   * */
-  const buildRequestDataIterate = (prompt) => {
-    // add IDs to the scenarios
-    scenarios.forEach((scenario, i) => {
-      if (scenario.id === undefined) scenario.id = i + 1;
-    });
-    return {
-      userinput: prompt,
-      scenarios: JSON.stringify(
-        scenarios
-          .filter((scenario) => scenario.exclude !== true)
-          .map(scenarioToJson),
-      ),
-      chatSessionId: chatSessionIdCardBuilding,
-    };
-  };
-
-  const iterateScenarios = (partiallyParsed) => {
-    partiallyParsed.forEach((parsedScenario) => {
-      const existingScenario = scenarios.find(
-        (scenario) => scenario.id === parsedScenario.id,
-      );
-      if (existingScenario) {
-        Object.assign(existingScenario, parsedScenario);
-        console.log(JSON.stringify(existingScenario));
-        // Remove any empty properties (sometimes happens with partial parsing)
-        Object.keys(existingScenario).forEach(
-          (key) =>
-            existingScenario[key] === "" ||
-            (existingScenario[key] === undefined &&
-              delete existingScenario[key]),
-        );
-      }
-    });
-    setScenarios([...scenarios]);
-  };
-
-  const sendIteration = (prompt) => {
-    setEnableGenerateMoreCards(false);
-    const uri = "/api/prompt/iterate";
-
-    let ms = "";
-    let output = [];
-
-    setIsGenerating(true);
-    setProgress(0);
-
-    // Simulate progress
-    const interval = setInterval(() => {
-      setProgress((prev) => {
-        if (prev >= 90) return prev; // avoid hitting 100 before done
-        return prev + 10;
-      });
-    }, 500);
-
-    fetchSSE(
-      uri,
-      {
-        method: "POST",
-        signal: startLoad(),
-        body: JSON.stringify(buildRequestDataIterate(prompt)),
-      },
-      {
-        json: true,
-        onErrorHandle: () => {
-          abortLoad();
-          clearInterval(interval);
-          setProgress(100);
-          setTimeout(() => {
-            setIsGenerating(false);
-            setProgress(0); // reset if you want
-          }, 1000);
-        },
-        onFinish: () => {
-          if (ms == "") {
-            toast.warning(
-              "Model failed to respond rightly, please rewrite your message and try again",
-            );
-          }
-          abortLoad();
-          clearInterval(interval);
-          setProgress(100);
-          setTimeout(() => {
-            setIsGenerating(false);
-            setProgress(0); // reset if you want
-          }, 1000);
-        },
-        onMessageHandle: (data) => {
-          if (data.data) {
-            ms += data.data;
-            ms = ms.trim().replace(/^[^[]+/, "");
-            if (ms.startsWith("[")) {
-              try {
-                output = parse(ms || "[]");
-              } catch (error) {
-                console.log("error", error);
-              }
-              if (Array.isArray(output)) {
-                iterateScenarios(output);
-              } else {
-                abortLoad();
-                if (ms.includes("Error code:")) {
-                  toast.error(ms);
-                } else {
-                  toast.warning(
-                    "Model failed to respond rightly, please rewrite your message and try again",
-                  );
-                }
-                console.log("response is not parseable into an array");
-              }
-            }
-          }
-        },
-      },
-    );
-  };
-
-  /** END ITERATION EXPERIMENT CODE */
 
   const sendFollowUpPrompt = (apiEndpoint, onData, followUpId) => {
     let ms = "";
@@ -671,10 +545,6 @@ const CardsChat = ({
     }
   };
 
-  const submitIterationPrompt = (prompt) => {
-    setIterationPrompt(prompt);
-  };
-
   return (
     <>
       <Drawer
@@ -722,51 +592,25 @@ const CardsChat = ({
                   </Button>
                 </div>
               )}
-              {/* Iteration experiment */}
-              {featureToggleConfig["cards_iteration"] === true &&
-                scenarios.length > 0 && (
-                  <div style={{ width: "88%", paddingLeft: "2em" }}>
-                    <h3>What else do you want to add to the cards?</h3>
-                    <Disclaimer message="Clicking 'Enrich Cards' will disable the 'Generate More Cards' button." />
-                    <HorizontalPossibilityPanel
-                      disable={loading}
-                      scenarioQueries={
-                        selectedPromptConfiguration.scenario_queries || []
-                      }
-                      onClick={submitIterationPrompt}
-                      tempClick={onFollowUp}
-                    />
-
-                    <div style={{ display: "flex", gap: "8px" }}>
-                      <Input
-                        value={iterationPrompt}
-                        onChange={(e, v) => {
-                          setIterationPrompt(e.target.value);
-                        }}
-                      />
-                      <Button
-                        style={{ marginBottom: "1px" }}
-                        disabled={loading}
-                        onClick={() => {
-                          setEnableGenerateMoreCards(false);
-                          sendIteration(iterationPrompt);
-                        }}
-                        className="go-button"
-                      >
-                        ENRICH CARDS
-                      </Button>
-                    </div>
-                  </div>
-                )}
-              {/* / End iteration experiment */}
+              <EnrichCard
+                startLoad={startLoad}
+                abortLoad={abortLoad}
+                loading={loading}
+                featureToggleConfig={featureToggleConfig}
+                chatSessionIdCardBuilding={chatSessionIdCardBuilding}
+                scenarios={scenarios}
+                setScenarios={setScenarios}
+                selectedPromptConfiguration={selectedPromptConfiguration}
+                setEnableGenerateMoreCards={setEnableGenerateMoreCards}
+                setIsGenerating={setIsGenerating}
+                setProgress={setProgress}
+                scenarioToJson={scenarioToJson}
+              />
               {scenarios.length > 0 && followUpCollapseItems.length > 0 && (
                 <div className="follow-up-container">
                   <div style={{ marginTop: "1em" }}>
                     <h3>What you can do next</h3>
-                    <p>
-                      Generate content based on the cards above. Remove cards to
-                      exclude them from the next step.
-                    </p>
+                    <p>Generate content based on the cards above.</p>
                   </div>
                   <Collapse
                     items={followUpCollapseItems}
